+++ ascenderkit/api/pages/access_list.py
@@ -12,7 +12,6 @@ page.register_page(
         resources.organization_access_list,
         resources.user_access_list,
         resources.inventory_access_list,
-        resources.group_access_list,
         resources.credential_access_list,
         resources.project_access_list,
         resources.job_template_access_list,