+++ ascenderkit/api/pages/settings.py
@@ -24,8 +24,6 @@ page.register_page(
         resources.settings_system,
         resources.settings_tacacsplus,
         resources.settings_ui,
-        resources.settings_user,
-        resources.settings_user_defaults,
     ],
     Setting,
 )